+++ src/darwin/mod.rs
@@ -2,6 +2,7 @@
 
 #![allow(improper_ctypes)]
 use std::collections::HashMap;
+use std::iter;
 use std::path::PathBuf;
 use std::ptr;
 
@@ -36,8 +37,13 @@ pub mod byte_order;
 use byte_order::kCGBitmapByteOrder32Host;
 
 use super::{
-    BitmapBuffer, FontDesc, FontKey, GlyphKey, Metrics, RasterizedGlyph, Size, Slant, Style, Weight,
+    BitmapBuffer, Error, FontDesc, FontKey, GlyphKey, Metrics, RasterizedGlyph, Size, Slant, Style,
+    Weight,
 };
+
+/// According to the documentation, the index of 0 must be a missing glyph character:
+/// https://developer.apple.com/fonts/TrueType-Reference-Manual/RM07/appendixB.html
+const MISSING_GLYPH_INDEX: u32 = 0;
 
 /// Font descriptor.
 ///
@@ -76,42 +82,7 @@ pub struct Rasterizer {
     use_thin_strokes: bool,
 }
 
-/// Errors occurring when using the core text rasterizer.
-#[derive(Debug)]
-pub enum Error {
-    /// Tried to rasterize a glyph but it was not available.
-    MissingGlyph(char),
-
-    /// Couldn't find font matching description.
-    MissingFont(FontDesc),
-
-    /// Requested an operation with a FontKey that isn't known to the rasterizer.
-    FontNotLoaded,
-}
-
-impl ::std::error::Error for Error {
-    fn description(&self) -> &str {
-        match *self {
-            Error::MissingGlyph(ref _c) => "Unable to find the requested glyph",
-            Error::MissingFont(ref _desc) => "Unable to find the requested font",
-            Error::FontNotLoaded => "Tried to operate on font that hasn't been loaded",
-        }
-    }
-}
-
-impl ::std::fmt::Display for Error {
-    fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
-        match *self {
-            Error::MissingGlyph(ref c) => write!(f, "Glyph not found for char {:?}", c),
-            Error::MissingFont(ref desc) => write!(f, "Unable to find the font {}", desc),
-            Error::FontNotLoaded => f.write_str("Tried to use a font that hasn't been loaded"),
-        }
-    }
-}
-
 impl crate::Rasterize for Rasterizer {
-    type Err = Error;
-
     fn new(device_pixel_ratio: f32, use_thin_strokes: bool) -> Result<Rasterizer, Error> {
         Ok(Rasterizer {
             fonts: HashMap::new(),
@@ -123,7 +94,7 @@ impl crate::Rasterize for Rasterizer {
 
     /// Get metrics for font specified by FontKey.
     fn metrics(&self, key: FontKey, _size: Size) -> Result<Metrics, Error> {
-        let font = self.fonts.get(&key).ok_or(Error::FontNotLoaded)?;
+        let font = self.fonts.get(&key).ok_or(Error::UnknownFontKey)?;
 
         Ok(font.metrics())
     }
@@ -144,20 +115,24 @@ impl crate::Rasterize for Rasterizer {
     /// Get rasterized glyph for given glyph key.
     fn get_glyph(&mut self, glyph: GlyphKey) -> Result<RasterizedGlyph, Error> {
         // Get loaded font.
-        let font = self.fonts.get(&glyph.font_key).ok_or(Error::FontNotLoaded)?;
+        let font = self.fonts.get(&glyph.font_key).ok_or(Error::UnknownFontKey)?;
 
-        // First try the font itself as a direct hit.
-        self.maybe_get_glyph(glyph, font).unwrap_or_else(|| {
-            // Then try fallbacks.
-            for fallback in &font.fallbacks {
-                if let Some(result) = self.maybe_get_glyph(glyph, &fallback) {
-                    // Found a fallback.
-                    return result;
-                }
-            }
-            // No fallback, give up.
-            Err(Error::MissingGlyph(glyph.c))
-        })
+        // Find a font where the given character is present.
+        let (font, glyph_index) = iter::once(font)
+            .chain(font.fallbacks.iter())
+            .find_map(|font| match font.glyph_index(glyph.character) {
+                MISSING_GLYPH_INDEX => None,
+                glyph_index => Some((font, glyph_index)),
+            })
+            .unwrap_or((font, MISSING_GLYPH_INDEX));
+
+        let glyph = font.get_glyph(glyph.character, glyph_index, self.use_thin_strokes);
+
+        if glyph_index == MISSING_GLYPH_INDEX {
+            Err(Error::MissingGlyph(glyph))
+        } else {
+            Ok(glyph)
+        }
     }
 
     fn update_dpr(&mut self, device_pixel_ratio: f32) {
@@ -182,7 +157,7 @@ impl Rasterizer {
             }
         }
 
-        Err(Error::MissingFont(desc.to_owned()))
+        Err(Error::FontNotFound(desc.to_owned()))
     }
 
     fn get_matching_face(
@@ -205,7 +180,7 @@ impl Rasterizer {
             }
         }
 
-        Err(Error::MissingFont(desc.to_owned()))
+        Err(Error::FontNotFound(desc.to_owned()))
     }
 
     fn get_font(&mut self, desc: &FontDesc, size: Size) -> Result<Font, Error> {
@@ -215,21 +190,6 @@ impl Rasterizer {
                 self.get_matching_face(desc, slant, weight, size)
             },
         }
-    }
-
-    /// Helper to try and get a glyph for a given font. Used for font fallback.
-    fn maybe_get_glyph(
-        &self,
-        glyph: GlyphKey,
-        font: &Font,
-    ) -> Option<Result<RasterizedGlyph, Error>> {
-        let scaled_size = self.device_pixel_ratio * glyph.size.as_f32_pts();
-        font.get_glyph(glyph.c, f64::from(scaled_size), self.use_thin_strokes)
-            .map(|r| Some(Ok(r)))
-            .unwrap_or_else(|e| match e {
-                Error::MissingGlyph(_) => None,
-                _ => Some(Err(e)),
-            })
     }
 }
 
@@ -438,7 +398,7 @@ impl Font {
     }
 
     fn glyph_advance(&self, character: char) -> f64 {
-        let index = self.glyph_index(character).unwrap();
+        let index = self.glyph_index(character);
 
         let indices = [index as CGGlyph];
 
@@ -455,12 +415,9 @@ impl Font {
     pub fn get_glyph(
         &self,
         character: char,
-        _size: f64,
+        glyph_index: u32,
         use_thin_strokes: bool,
-    ) -> Result<RasterizedGlyph, Error> {
-        let glyph_index =
-            self.glyph_index(character).ok_or_else(|| Error::MissingGlyph(character))?;
-
+    ) -> RasterizedGlyph {
         let bounds = self.bounding_rect_for_glyph(Default::default(), glyph_index);
 
         let rasterized_left = bounds.origin.x.floor() as i32;
@@ -471,14 +428,14 @@ impl Font {
         let rasterized_height = (rasterized_descent + rasterized_ascent) as u32;
 
         if rasterized_width == 0 || rasterized_height == 0 {
-            return Ok(RasterizedGlyph {
-                c: ' ',
+            return RasterizedGlyph {
+                character: ' ',
                 width: 0,
                 height: 0,
                 top: 0,
                 left: 0,
-                buf: BitmapBuffer::RGB(Vec::new()),
-            });
+                buffer: BitmapBuffer::RGB(Vec::new()),
+            };
         }
 
         let mut cg_context = CGContext::create_bitmap_context(
@@ -530,31 +487,31 @@ impl Font {
 
         let rasterized_pixels = cg_context.data().to_vec();
 
-        let buf = if is_colored {
+        let buffer = if is_colored {
             BitmapBuffer::RGBA(byte_order::extract_rgba(&rasterized_pixels))
         } else {
             BitmapBuffer::RGB(byte_order::extract_rgb(&rasterized_pixels))
         };
 
-        Ok(RasterizedGlyph {
-            c: character,
+        RasterizedGlyph {
+            character,
             left: rasterized_left,
             top: (bounds.size.height + bounds.origin.y).ceil() as i32,
             width: rasterized_width as i32,
             height: rasterized_height as i32,
-            buf,
-        })
+            buffer,
+        }
     }
 
-    fn glyph_index(&self, character: char) -> Option<u32> {
+    fn glyph_index(&self, character: char) -> u32 {
         // Encode this char as utf-16.
-        let mut buf = [0; 2];
-        let encoded: &[u16] = character.encode_utf16(&mut buf);
+        let mut buffer = [0; 2];
+        let encoded: &[u16] = character.encode_utf16(&mut buffer);
         // And use the utf-16 buffer to get the index.
         self.glyph_index_utf16(encoded)
     }
 
-    fn glyph_index_utf16(&self, encoded: &[u16]) -> Option<u32> {
+    fn glyph_index_utf16(&self, encoded: &[u16]) -> u32 {
         // Output buffer for the glyph. for non-BMP glyphs, like
         // emojis, this will be filled with two chars the second
         // always being a 0.
@@ -569,9 +526,9 @@ impl Font {
         };
 
         if res {
-            Some(u32::from(glyphs[0]))
+            u32::from(glyphs[0])
         } else {
-            None
+            MISSING_GLYPH_INDEX
         }
     }
 }
@@ -598,19 +555,19 @@ mod tests {
 
         for font in fonts {
             // Get a glyph.
-            for c in &['a', 'b', 'c', 'd'] {
-                let glyph = font.get_glyph(*c, 72., false).unwrap();
+            for character in &['a', 'b', 'c', 'd'] {
+                let glyph_index = font.glyph_index(*character);
+                let glyph = font.get_glyph(*character, glyph_index, false);
 
-                let buf = match &glyph.buf {
-                    BitmapBuffer::RGB(buf) => buf,
-                    BitmapBuffer::RGBA(buf) => buf,
+                let buffer = match &glyph.buffer {
+                    BitmapBuffer::RGB(buffer) | BitmapBuffer::RGBA(buffer) => buffer,
                 };
 
                 // Debug the glyph.. sigh.
                 for row in 0..glyph.height {
                     for col in 0..glyph.width {
                         let index = ((glyph.width * 3 * row) + (col * 3)) as usize;
-                        let value = buf[index];
+                        let value = buffer[index];
                         let c = match value {
                             0..=50 => ' ',
                             51..=100 => '.',